Add files via upload
This project was made by a team during Global AI HUB Deep Learning bootcamp. I would like to thank my teammates İlker, Serdar and Tuna for their contributions.
It was quite nice to experience doing a project with a team.Thank you to the entire Global AI Hub team for creating this environment for us.